Program for creating token loggers, feel free to use
All you do is run the program; it will pop up with a window, which you type in the web hook of your own bot, add a custom image or link, and it will take it from there! It will copy the link to your clipboard and all you have to do is send it to your victim! Please, only use this on people who have allowed you to do this, this is intended to be used for making protections against programs like this.